Removed the general package and moved everything to the main package
