+++ src/main/java/me/partlysanestudios/partlysaneskies/Main.java
@@ -10,29 +10,25 @@ import java.util.List;
 import org.apache.commons.lang3.reflect.FieldUtils;
 
 import gg.essential.elementa.ElementaVersion;
+import me.partlysanestudios.partlysaneskies.auctionhouse.AhManager;
+import me.partlysanestudios.partlysaneskies.chatalerts.ChatAlertsCommand;
+import me.partlysanestudios.partlysaneskies.chatalerts.ChatAlertsManager;
 import me.partlysanestudios.partlysaneskies.dungeons.WatcherReady;
 import me.partlysanestudios.partlysaneskies.dungeons.partymanager.PartyManager;
 import me.partlysanestudios.partlysaneskies.dungeons.partymanager.PartyManagerCommand;
 import me.partlysanestudios.partlysaneskies.dungeons.permpartyselector.PermPartyCommand;
 import me.partlysanestudios.partlysaneskies.dungeons.permpartyselector.PermPartyManager;
 import me.partlysanestudios.partlysaneskies.garden.GardenTradeValue;
-import me.partlysanestudios.partlysaneskies.general.LocationBannerDisplay;
-import me.partlysanestudios.partlysaneskies.general.NoCookieWarning;
-import me.partlysanestudios.partlysaneskies.general.WikiArticleOpener;
-import me.partlysanestudios.partlysaneskies.general.WormWarning;
-import me.partlysanestudios.partlysaneskies.general.chatalerts.ChatAlertsCommand;
-import me.partlysanestudios.partlysaneskies.general.chatalerts.ChatAlertsManager;
-import me.partlysanestudios.partlysaneskies.general.economy.auctionhouse.AhManager;
-import me.partlysanestudios.partlysaneskies.general.partyfriend.PartyFriendManager;
-import me.partlysanestudios.partlysaneskies.general.partyfriend.PartyFriendManagerCommand;
-import me.partlysanestudios.partlysaneskies.general.petalert.PetAlert;
-import me.partlysanestudios.partlysaneskies.general.petalert.PetAlertMuteCommand;
-import me.partlysanestudios.partlysaneskies.general.rngdropbanner.DropBannerDisplay;
-import me.partlysanestudios.partlysaneskies.general.skillupgrade.SkillUpgradeCommand;
-import me.partlysanestudios.partlysaneskies.general.skillupgrade.SkillUpgradeRecommendation;
 import me.partlysanestudios.partlysaneskies.help.ConfigCommand;
 import me.partlysanestudios.partlysaneskies.help.DiscordCommand;
 import me.partlysanestudios.partlysaneskies.help.HelpCommand;
+import me.partlysanestudios.partlysaneskies.partyfriend.PartyFriendManager;
+import me.partlysanestudios.partlysaneskies.partyfriend.PartyFriendManagerCommand;
+import me.partlysanestudios.partlysaneskies.petalert.PetAlert;
+import me.partlysanestudios.partlysaneskies.petalert.PetAlertMuteCommand;
+import me.partlysanestudios.partlysaneskies.rngdropbanner.DropBannerDisplay;
+import me.partlysanestudios.partlysaneskies.skillupgrade.SkillUpgradeCommand;
+import me.partlysanestudios.partlysaneskies.skillupgrade.SkillUpgradeRecommendation;
 import me.partlysanestudios.partlysaneskies.utils.Utils;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiScreen;
